feat: supplementary string names the comments for each function
