+++ src/commands/string_commands/Cargo.toml
@@ -7,6 +7,7 @@ edition = "2021"
 
 [dependencies]
 bigdecimal = "0.4.2"
+lazy_static = "1.4.0"
 
 # db crate
 rus_key_db = { path = "../../db" }
